+++ R/Dates.r
@@ -26,3 +26,20 @@ Date <- function(length = 0L) {
 #' @rdname date_utils
 #' @export
 NA_Date_ <- structure(NA_real_, class = "Date")
+
+.recursive_date_unclass <- function(x) {
+  if (is.recursive(x))
+    lapply(x, .recursive_date_unclass)
+  else
+    as_date(x)
+}
+
+#' @method c Date
+c.Date <- function(..., recursive = FALSE) {
+  structure(c(unlist(lapply(list(...), .recursive_date_unclass))),
+            class = "Date")
+}
+
+evalqOnLoad({
+    registerS3method("c", "Date", c.Date)
+})